requirements.txt
picamera
opencv-python
numpy
pandas
pyserial

# sudo apt install -y python3-libcamera
# pip3 install picamera2 --upgrade